Update stanfit.R
This is less of bug fix and more of a 'best practice'.  We need to set the seed within the call of stan.  Any parameters for which initial values were not specified will receive a randomly generated initial value.  Using R's set.seed function to set the seed for Stan will not work per the documentation.